+++ src/app/lib/prisma.ts
@@ -1,6 +1,10 @@
 import { PrismaClient } from "@prisma/client";
 
-let prisma;
+declare global {
+  var prisma: PrismaClient;
+}
+
+let prisma: PrismaClient;
 
 if (process.env.NODE_ENV === "production") {
   prisma = new PrismaClient();
